+++ src/conv.rs
@@ -3,8 +3,8 @@
 
 mod from_ocaml;
 mod to_ocaml;
-mod into_rust;
+mod to_rust;
 
 pub use self::from_ocaml::FromOCaml;
 pub use self::to_ocaml::ToOCaml;
-pub use self::into_rust::IntoRust;
+pub use self::to_rust::ToRust;