+++ src/conv/to_rust.rs
@@ -5,16 +5,16 @@ use crate::conv::FromOCaml;
 use crate::value::OCaml;
 
 /// Counterpart to [`FromOCaml`], usually more convenient to use.
-pub trait IntoRust<T>: Sized {
+pub trait ToRust<T>: Sized {
     /// Convert into a Rust value.
-    fn into_rust(self) -> T;
+    fn to_rust(self) -> T;
 }
 
-impl<'a, OCamlT, RustT> IntoRust<RustT> for OCaml<'a, OCamlT>
+impl<'a, OCamlT, RustT> ToRust<RustT> for OCaml<'a, OCamlT>
 where
     RustT: FromOCaml<OCamlT>,
 {
-    fn into_rust(self) -> RustT {
+    fn to_rust(self) -> RustT {
         RustT::from_ocaml(self)
     }
 }